pyproject.toml
[project]
name = "advanced-rag"
version = "0.1.0"
description = "Advanced multimodal RAG over technical documents with LangChain, LangGraph, LlamaIndex, Weaviate, Neo4j, and ColPali."
requires-python = ">=3.13"
dependencies = [
    "colpali-engine>=0.3.15",
    "fastapi>=0.136.0",
    "flagembedding>=1.3.5",
    "langchain>=1.2.15",
    "langchain-cohere>=0.5.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-neo4j>=0.9.0",
    "langchain-openai>=1.1.14",
    "langchain-voyageai>=0.3.3",
    "langchain-weaviate>=0.0.6",
    "langgraph>=1.1.8",
    "langsmith>=0.7.32",
    "llama-index>=0.14.20",
    "llama-index-embeddings-huggingface>=0.7.0",
    "llama-index-embeddings-openai>=0.6.0",
    "llama-index-embeddings-voyageai>=0.6.0",
    "llama-index-graph-stores-neo4j>=0.7.0",
    "llama-index-llms-openai>=0.7.5",
    "llama-index-vector-stores-weaviate>=1.6.0",
    "llama-parse>=0.6.94",
    "pillow>=12.2.0",
    "python-multipart>=0.0.26",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "streamlit>=1.56.0",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/advanced_rag"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
